# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ToolHelper local service contract",
    "version": "0.1.0"
  },
  "servers": [{ "url": "http://127.0.0.1" }],
  "paths": {
    "/actuator/health": {
      "get": {
        "summary": "Java service health",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/HealthResponse" } }
      }
    },
    "/health": {
      "get": {
        "summary": "C# local service health",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/HealthResponse" } }
      }
    },
    "/api/jobs/{jobId}/events": {
      "get": {
        "summary": "Replayable job event stream",
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "Last-Event-ID", "in": "header", "required": false, "schema": { "type": "string" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "SSE stream", "content": { "text/event-stream": {} } } }
      }
    },
    "/api/database/sessions": {
      "post": {
        "summary": "Open an isolated user SQLite session",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateSessionRequest" } } } },
        "responses": { "200": { "description": "Session opened", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
    "/api/database/sessions/{sessionId}/metadata": {
      "get": {
        "summary": "List SQLite objects",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "sessionId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Metadata", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
    "/api/database/sessions/{sessionId}/queries": {
      "post": {
        "summary": "Submit paged SQLite SQL",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "sessionId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Task accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
    "/api/database/queries/{taskId}/events": {
      "get": {
        "summary": "Replayable database query events",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "Last-Event-ID", "in": "header", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "SSE stream", "content": { "text/event-stream": {} } } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "ToolHelper service token" }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": ["success", "code", "message", "traceId"],
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": {},
          "traceId": { "type": "string" }
        }
      },
      "HealthData": {
        "type": "object",
        "required": ["service", "version", "status", "traceId"],
        "properties": {
          "service": { "type": "string" },
          "version": { "type": "string" },
          "status": { "type": "string", "enum": ["UP"] },
          "traceId": { "type": "string" }
        }
      },
      "CreateSessionRequest": {
        "type": "object",
        "required": ["path"],
        "properties": {
          "path": { "type": "string" },
          "password": { "type": "array", "items": { "type": "integer", "minimum": 0, "maximum": 65535 } }
        }
      },
      "SseEvent": {
        "type": "object",
        "required": ["eventId", "type", "jobId", "timestamp", "traceId", "payload"],
        "properties": {
          "eventId": { "type": "string" },
          "type": { "type": "string" },
          "jobId": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "traceId": { "type": "string" },
          "payload": { "type": "object" }
        }
      }
    },
    "responses": {
      "HealthResponse": {
        "description": "Health response",
        "headers": { "X-Trace-Id": { "schema": { "type": "string" } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } }
      }
    }
  }
}